more dev niceties

+++ tests/integration/restricted/test.sh
@@ -1,4 +1,9 @@
 #!/usr/bin/env bash
+#
+# Run the integration test that deploys a Sourcegraph cluster with a restrictive security policy.
+# Conveniently, this can be run from dev by setting the TEST_GCP_{ZONE,PROJECT,USERNAME} environment
+# variables (assuming your local `gcloud` is auth'd with the GCP username). Optionally set NOCLEANUP
+# to prevent cleaning up the cluster when finished.
 
 set -xeuo
 
@@ -23,8 +28,10 @@ cd $(dirname "${BASH_SOURCE[0]}")
 gcloud container clusters create ${CLUSTER_NAME} --zone ${TEST_GCP_ZONE} --num-nodes 3 --machine-type n1-standard-16 --disk-type pd-ssd --project ${TEST_GCP_PROJECT} --labels="cost-category=build,build-creator=${BUILD_CREATOR},build-branch=${BUILD_BRANCH},integration-test=fresh"
 
 gcloud container clusters get-credentials ${CLUSTER_NAME} --zone ${TEST_GCP_ZONE} --project ${TEST_GCP_PROJECT}
-CLUSTER_CLEANUP="gcloud container clusters delete ${CLUSTER_NAME} --zone ${TEST_GCP_ZONE} --project ${TEST_GCP_PROJECT} --quiet"
-CLEANUP="$CLUSTER_CLEANUP; $CLEANUP"
+if [ -z "${NOCLEANUP:-}" ]; then
+    CLUSTER_CLEANUP="gcloud container clusters delete ${CLUSTER_NAME} --zone ${TEST_GCP_ZONE} --project ${TEST_GCP_PROJECT} --quiet"
+    CLEANUP="$CLUSTER_CLEANUP; $CLEANUP"
+fi
 
 kubectl create clusterrolebinding cluster-admin-binding --clusterrole cluster-admin --user ${TEST_GCP_USERNAME}
